+++ lib/screens/setup/setup_auth.dart
@@ -115,9 +115,6 @@ class _SetupAuthScreenState extends State<SetupAuthScreen> {
                                   ),
                                 ),
                                 value: _biometricsAllowed,
-                                activeColor:
-                                    Theme.of(context).colorScheme.background,
-                                inactiveThumbColor: Colors.grey,
                                 onChanged: (newState) {
                                   if (_biometricsAvailable == false) {
                                     ScaffoldMessenger.of(context).showSnackBar(