+++ lib/widgets/wallet/address_book/addresses_tab_watch_only.dart
@@ -152,7 +152,7 @@ class _AddressesTabWatchOnlyState extends State<AddressesTabWatchOnly> {
                   formKey.currentState!.save();
                   context.read<WalletProvider>().createWatchOnlyAddres(
                         identifier: widget.walletName,
-                        address: addressController.text,
+                        address: addressController.text.trim(),
                         label: labelController.text == ''
                             ? ''
                             : labelController.text,
@@ -189,10 +189,8 @@ class _AddressesTabWatchOnlyState extends State<AddressesTabWatchOnly> {
                     padding: const EdgeInsets.all(20),
                     child: ElevatedButton(
                       style: ElevatedButton.styleFrom(
-                        foregroundColor:
-                            Theme.of(context).colorScheme.background,
-                        backgroundColor:
-                            Theme.of(context).colorScheme.background,
+                        foregroundColor: Theme.of(context).colorScheme.surface,
+                        backgroundColor: Theme.of(context).colorScheme.surface,
                         fixedSize: Size(
                           MediaQuery.of(context).size.width > 1200
                               ? MediaQuery.of(context).size.width / 5
@@ -239,7 +237,7 @@ class _AddressesTabWatchOnlyState extends State<AddressesTabWatchOnly> {
                         style: TextStyle(
                           fontSize: 16,
                           fontStyle: FontStyle.italic,
-                          color: Theme.of(context).colorScheme.background,
+                          color: Theme.of(context).colorScheme.surface,
                         ),
                       ),
                     ),