+++ lib/widgets/wallet/wallet_home_connection.dart
@@ -15,7 +15,7 @@ class WalletHomeConnection extends StatelessWidget {
       widget = Text(
         AppLocalizations.instance.translate('wallet_connected'),
         style: TextStyle(
-          color: Theme.of(context).colorScheme.background,
+          color: Theme.of(context).colorScheme.surface,
           letterSpacing: 1.4,
           fontSize: 16,
         ),
@@ -24,7 +24,7 @@ class WalletHomeConnection extends StatelessWidget {
       widget = Text(
         AppLocalizations.instance.translate('wallet_offline'),
         style: TextStyle(
-          color: Theme.of(context).colorScheme.background,
+          color: Theme.of(context).colorScheme.surface,
           fontSize: 16,
           letterSpacing: 1.4,
         ),